# openapi.yaml
openapi: 3.0.3
info:
  title: HRX API
  description: |
    API para gerenciamento de profissionais, fornecedores e projetos de eventos.

    ## Autenticação
    Todas as rotas requerem autenticação via Clerk. O token JWT deve ser enviado no header Authorization.

    ## Rate Limiting
    - Leitura: 100 requests/minuto
    - Escrita: 20 requests/minuto

    ## Otimizações
    - Índices geográficos para buscas por localização
    - RPC functions para queries complexas
    - Cache com Redis para dados frequentes
  version: 1.0.0
  contact:
    name: HRX Tecnologia
    email: contato@hrxeventos.com.br
  license:
    name: Proprietary
servers:
  - url: https://hrx.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT do Clerk
...
paths:
...
  /admin/geocode/batch:
    post:
      summary: Geocodificação em lote
      description: |
        Geocodifica múltiplos profissionais/fornecedores de uma vez.

        **Otimização:** Fix N+1 query (6x mais rápido - 60s → 10s para 100 itens)
      tags:
        - Geocoding
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [professionals, suppliers, events]
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
              required:
                - type
                - ids
      responses:
        '200':
          description: Geocodificação concluída
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            success:
                              type: boolean
                            latitude:
                              type: number
                            longitude:
                              type: number
                            formattedAddress:
                              type: string
                            error:
                              type: string
                      summary:
                        type: object
                        properties:
                          total:
                            type: integer
                          successful:
                            type: integer
                          failed:
                            type: integer
  /admin/map-data:
    get:
      summary: Dados para mapa
      description: |
        Retorna marcadores para visualização em mapa.

        **Otimizações:**
        - Suporte a viewport (bounding box)
        - Queries paralelas
        - Filtro por tipos de marcadores
      tags:
        - Map Data
      security:
        - ClerkAuth: []
      parameters:
        - name: minLat
          in: query
          schema:
            type: number
          description: Latitude mínima do viewport
        - name: maxLat
          in: query
          schema:
            type: number
        - name: minLng
          in: query
          schema:
            type: number
        - name: maxLng
          in: query
          schema:
            type: number
        - name: types
          in: query
          schema:
            type: string
          description: Tipos separados por vírgula (professional,supplier,event)
      responses:
        '200':
          description: Marcadores do mapa
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  markers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        type:
                          type: string
                          enum: [professional, supplier, event]
                        latitude:
                          type: number
                        longitude:
                          type: number
                        name:
                          type: string
                        city:
                          type: string
                        state:
                          type: string
                        status:
                          type: string
                  total:
                    type: integer
                  viewport:
                    type: object
                    nullable: true
                    properties:
                      minLat:
                        type: number
                      maxLat:
                        type: number
                      minLng:
                        type: number
                      maxLng:
                        type: number
security:
  - ClerkAuth: []